third_party/blink/web_tests/platform/generic/fast/css/fontfaceset-multiple-faces-expected.txt

Test load events for a font family consists of multiple @font-faces.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.fontfaces.length is 2
PASS event.fontfaces[0].status is "loaded"
PASS event.fontfaces[1].status is "loaded"
PASS document.fonts.check('10px TestFont') is true
PASS firedEvents is ['loading', 'loadingdone']
PASS successfullyParsed is true

TEST COMPLETE
